Fourth Set of Collected Bug Fixes and Maintenance Updates for 29 August 2024 Stable Release

+++ src/create_bonds.cpp
@@ -241,7 +241,7 @@ void CreateBonds::many()
   int *num_bond = atom->num_bond;
   int **bond_type = atom->bond_type;
   tagint **bond_atom = atom->bond_atom;
-  double newton_bond = force->newton_bond;
+  int newton_bond = force->newton_bond;
   int nlocal = atom->nlocal;
 
   int i, j, ii, jj, inum, jnum, flag;
